requirements.txt
email-validator==1.1.2
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-JSON==0.3.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Keras==2.4.3
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.4
plotly-express==0.4.1
scikit-learn==0.24.2
sklearn==0.0
SQLAlchemy==1.4.17
WTForms==2.3.3
yfinance==0.1.63